delint and fix
